# swagger.yaml
openapi: 3.0.0
info:
  title: Webser2 API
  description: Web servisi analitik ve izleme API'si
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your API key as the Bearer token
  schemas:
...
    AnalyticsResponse:
      type: object
      properties:
        totalRequests:
          type: array
          items:
            type: object
            properties:
              count:
                type: integer
                example: 1000
        requestsByEndpoint:
          type: array
          items:
            type: object
            properties:
              endpoint:
                type: string
                example: "/api/users"
              count:
                type: integer
                example: 500
        requestsByStatus:
          type: array
          items:
            type: object
            properties:
              status_code:
                type: integer
                example: 200
              count:
                type: integer
                example: 800
        averageResponseTime:
          type: array
          items:
            type: object
            properties:
              avg_response_time:
                type: number
                format: float
                example: 150.5
        errorRates:
          type: array
          items:
            type: object
            properties:
              endpoint:
                type: string
                example: "/api/users"
              total_requests:
                type: integer
                example: 1000
              error_count:
                type: integer
                example: 50
              error_rate:
                type: number
                format: float
                example: 5.00
...
security:
  - BearerAuth: []
paths:
...
  /api/analytics:
    get:
      summary: Analitik verilerini getir
      description: Belirtilen tarih aralığı için analitik verilerini döndürür
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: range
          schema:
            type: string
            enum: [today, week, month]
          description: Önceden tanımlanmış zaman aralığı
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Özel tarih aralığı başlangıcı
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: Özel tarih aralığı sonu
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '400':
          description: Geçersiz istek
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid date range"
        '500':
          description: Sunucu hatası
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
...
  /api/job/result/{jobId}:
    get:
      summary: Download job results
      description: Download the query results as a ZIP file containing JSON
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: Job identifier
      responses:
        '200':
          description: Results downloaded successfully
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/database:
    get:
      summary: List all tenants and their database IDs
      description: Returns a list of all tenants and their corresponding database IDs
      responses:
        '200':
          description: Successfully retrieved all tenant database IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    tenantId:
                      type: string
                      example: "tavukdunyasi"
                    databaseId:
                      type: string
                      example: "3"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "An unexpected error occurred"
                  details:
                    type: object
                    properties:
                      server:
                        type: string
                      database:
                        type: string
                      user:
                        type: string
  /api/database/{tenantId}:
    get:
      summary: Get database ID for a tenant or list all tenants
      description: Returns the database ID for a specific tenant. Use "all" or leave empty to get all tenants and their database IDs.
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: string
          description: The ID of the tenant (e.g., tavukdunyasi, kahvedunyasi, demo) or "all" to list all tenants
          example: tavukdunyasi
      responses:
        '200':
          description: Successfully retrieved database ID(s)
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      databaseId:
                        type: string
                        example: "3"
                  - type: array
                    items:
                      type: object
                      properties:
                        tenantId:
                          type: string
                          example: "tavukdunyasi"
                        databaseId:
                          type: string
                          example: "3"
        '404':
          description: Tenant ID not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Tenant ID 'invalid' not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "An unexpected error occurred"
                  details:
                    type: object
                    properties:
                      server:
                        type: string
                      database:
                        type: string
                      user:
                        type: string